requirements.txt
loguru==0.7.3
pillow==11.1.0
pysdl2==0.9.17
pydantic==1.10